requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
appdirs==1.4.4
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
enum34==1.1.10
filelock==3.0.12
idna==2.10
iniconfig==1.0.1
lxml==4.5.2
more-itertools==8.4.0
namedlist==1.7
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
requests==2.25.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.11
virtualenv==20.0.30
